Make camera positionable

+++ src/lib/Utils.h
@@ -4,7 +4,7 @@
 
 namespace Raytracing
 {
-    inline double random()
+    inline float random()
     {
         static thread_local std::uniform_real_distribution<float> distribution(0.0F, 1.0F);
         static thread_local std::mt19937 generator;
